Add middleware in redux store for async code
