+++ src/redux/configureStore.js
@@ -1,6 +1,9 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
+import thunk from 'redux-thunk';
 import booksReducer from './books/books';
+
+const middlewareList = [thunk, logger];
 
 const reducer = combineReducers({
   booksReducer,
@@ -9,7 +12,7 @@ const reducer = combineReducers({
 
 const store = createStore(
   reducer,
-  applyMiddleware(logger),
+  applyMiddleware(...middlewareList),
 );
 
 export default store;
